Include group types in the custom show package schema
The group item links in the group membership view for a dataset (ie. the
path /dataset/groups/<dataset>) were not generated for the appropriate
group types, because that information wasn't included in the datasets
show package schema (so wasn't available in the template context when
generating links).

This change appends group types to the show package schema.